requirements.txt
# SensePedia 自动化测试框架 - 依赖项

# 核心依赖
httpx>=0.25.0          # HTTP 客户端
pyyaml>=6.0            # YAML 配置解析
pydantic>=2.0.0        # 数据验证（可选）

# Excel 处理
openpyxl>=3.1.0        # Excel 文件读取

# 异步支持
anyio>=4.0.0           # 异步网络 I/O

# 测试框架（可选，用于单元测试）
pytest>=7.4.0
pytest-asyncio>=0.21.0
